nvidia-drm: convert &NvKmsKapiFunctionsTable to direct calls
Same as with &nvidia_modeset_rm_ops_t, no reason to have those as
callbacks. Okay, they come from the "standalone" src/nvidia-modeset,
but that doesn't have to mean anything, at least now that it is just
a pile of sources for Kbuild.

So, make it the following way:
- src/nvidia-modeset now has a bunch of "interface" functions. They
  are global and wrap former static callbacks, so the latters get
  inlined into the formers by the compiler;
- nvidia-modeset.ko export them as regular interface symbols;
- nvidia-drm uses those exported functions directly.

This rips off several Kb from the affected modules, always nice to
have. Remove ::getFlipPendingStatus() at all as it's not used
anywhere.

Signed-off-by: Alexander Lobakin <alobakin@pm.me>