treewide: fix most of -Wmissing-prototypes
Enabled on W=1, the warning allows to find at least some of the
functions, which can be made static. The rest needs to add some
missing includes.
This greatly optimizes "lr10"-related code as at the end it is
just a callback table, most of which are declared and used only
within one file.

Signed-off-by: Alexander Lobakin <alobakin@pm.me>