treewide: streamline a ton of externs and open-code
Lots of different Linux function were wrapped into oneliners to
ensure it will link correctly with closed binaries and/or CXX
objects, plus to provide common prototypes between different OSes.
As this particular fork is for the Linux kernel only, and for the
latest stable only, unwrap most of them with converting to a static
inline, which calls the original Linux function, or just define a
macro where applicable. This gives way better optimization results
and far better diagnostics coverage, as plenty of arguments are
always compile-time constant, but the compiler can't track them when
they're behind the external call.
There's a couple of externs left in nvidia-modeset. They're prefixed
with __ and they are there to be called from the CXX code, as it
can't properly inline C headers.

Signed-off-by: Alexander Lobakin <alobakin@pm.me>