nvidia: optimize LR10 discovery process
It contains some huge tables/arrays on the stack, but the main
problem, which even makes the compilers compile the file for more
than a minute, is that one macro at the end of the file repeats
the same code for 23+ times, and it can't be uninlined or folded
into smaller chunks.
Use offsetof() to be able to associate structure members with enums
or strings instead of generating those tables at runtime, and
convert that 23+ macro call into a loop over an array with data.
This wipes more than 25 Kb from the text size, nice.

Signed-off-by: Alexander Lobakin <alobakin@pm.me>