conftest: make it a read-only checker
This idea of generating header files on the go could be not so bad
if, for example, conftest didn't need Module.symvers to work. With
this little detail, it was impossible to just integrate it into the
tree and watch it build -- you needed a first pass completed before
enabling the driver. Not speaking of all those source files waiting
for the single pipe to process a particular conftest header.
So, do it just like it's done in the kernel with the atomic headers.
Provide already generated and stable conftest headers and make it
buildable right from the start, but still run conftest to ensure
that those headers are up-to-date. If no, Kbuild will then trigger
an error and interrupt the build. This is needed in order to keep
them actual between the driver updates -- conftests is being changed
quite often, so letting them rot for even a bit means potential
build or runtime problems.
The rules are not mature enough, to be rewritten later.

Signed-off-by: Alexander Lobakin <alobakin@pm.me>