treewide: deduplicate common code into nvidia-common
Create a new module and move nvstatus.c and nv-kthread-q.c to it.
The first one is used by both nvidia and nvidia-modeset, which is
unsafe (will be prohibited soon by Kbuild), also nvidia-uvm has a
copy of it in its folder. The second is copied over pretty much
each module with no reason.
Dedup that functionality and provide it with the new nvidia-common.
There'll be more stuff moved to it possibly later on, for now that
was the most obvious stuff.

Signed-off-by: Alexander Lobakin <alobakin@pm.me>