treewide: finish fixing -Wmissing-prototypes -Wunused-macros
The second one is rather silly, but sometimes allows to find piles
of forsaken stuff.
The first is a must, and Nvidia is bad at it. I was fixing it
already, but that was only a part of all warning. This time there
are no more warnings at all and nvidia.ko got -18 Kb off its object
code size.

Signed-off-by: Alexander Lobakin <alobakin@pm.me>