shared: fix Callback5ArgVoidReturn CFI violation
The mentioned type has @hEvent defined as NvHandle and the last one,
@status, as NvU32. However, different callback implementations across
the shared code (mostly in nvidia-modeset) were defined differently:
either @hEvent was NvU32 or @status was NV_STATUS. Surprisingly enough,
Clang 16 (the latest Git snapshot) was not able to catch this, despite
that NV_STATUS is an enum. Plus my machine was either hanging silently
or throwing an "invalid opcode: 0000" panic, so it took a bunch to
realize what's wrong here.
Fix this by making @status of type NV_STATUS and adjusting all the
implementations across the code to that definiton (void *, void *,
NvEvent, NvU32, NV_STATUS). Now the driver seems to work for the regular
daily routines...

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>