shared: replace all of strncpy()s with strscpy()
strncpy() is generally deprecated and should not be used in new code.
Hovewer, they're a bit different with strscpy() (the recommended
replacement) in terms of the return value type: the former returns the
first argument (with no modifications), where the latter returns the
number of bytes copied (or -%ETOOBIG).
NVIDIA doesn't check return values anywhere, so the replacement is
straightforward and safe.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>