shared: nvidia: use Linux atomics instead of C11
C11 compiler-builtin atomics are incompatible with the Linux kernel
memory model and shouldn't be used here. This was discussed several
times on LKML. I know the intention was to use environment-independent
implementation, but this is not how Linux drivers work.
Reimplement nvport/atomics.h using the Linux atomics operations.
Unfortunately, this can't be done without changing the types of all the
variables and fields used atomically. Sure, atomic_t has the same size
as NvS32 has and so on, so the first implementation was simply casting
them with checking for size. However, there are more than just a few of
places where Nvidia uses atomic objects as just regular ones.
For example, there was no macro for reading values, as they were read
as just simple direct fetch. This might work, but might not work as well
depending on some luck and codegen. So, the solution without any casting
and with using the proper types makes sure there are no such misuses
left.
Atomic types are mapped as follows:

NvS32  -> atomic_t
NvU32  -> atomic_t
NvS64  -> atomic64_t
NvU64  -> atomic64_t
NvSPtr -> atomic_long_t

This doesn't change the size of any structure since the types are fully
size-compatible. The only change done is the layout of
`struct PORT_MEM_COUNTER`: the 4-byte hole can be avoided by placing
wider types before the 4-byte ones.
In one place, atomic op was replaced with xchg(), as the intention there
was to atomically swap two pointers, but the atomic-long op was used to
do that ._.

Initially, this problem was spotted on BTF-enabled builds: pahole's BTF
encoder doesn't support _Atomic, which is used in C11 implementations.
Now that they're gone, limiting the debuginfo level to 1 for nvidia.ko
is not needed, so remove it.
Limiting the C++ code debug level is still needed tho due to that pahole
doesn't support C++' references. This might change once full Rust
support is here, but for now reword the comment next to it a bit.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>