treewide: convert all fake 0- and 1-element arrays into flexible arrays
Zero- and one-sized arrays to represent variable length arrays are
deprecated, error-prone and are not digestable by modern compilers.
They can track dynamic allocations, but not when there was some fake
array involved. So, -fstrict-flex-arrays=3 is inevitable to appear in
the kernel sooner-later, and while it's on its way, fix all the
instances in the NVIDIA driver.
This involves a lot, most notably there are no more "onstack memory
descriptors", as having a variable-length array on the stack is
something... not really correct. uvm_tracker_t is left as it is for
now, but it's beyond ugly TBH and may get properized soon as well. Just
too many instances of it that need careful medication and testing.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>